fix(llm): sanitize import warning messages
